[Adjustment][Maxell] refactored fb post component
