+++ src/constants/FbApiEndpoints.ts
@@ -1,0 +1,2 @@
+export const pageRequestEndpoint = `${process.env.URL}/api/socmed/facebook`;
+export const latestPostRequestEndpoint = `${process.env.URL}/api/socmed/facebook/latest-post`;
